Update Feature-based Molecular Network input .xml
Correction and updates provided here. What documentation link should be provided in "Precursor Ion Mass tolerance" description in the tool box? (one to the "(featurebasedmolecularnetworking)"?)
There is no tool box for quantification_table and metadata_table in the File Selection box of the workflow. Will be done in the next update.