Allowing and removing duplicated filename
Instead of breaking the FBMN workflow if there are duplicated filename in the OpenMS table. We allow duplicated filenames and keep only the first occurence.
It is not ideal, but duplicated filenames are often for low importances samples like Blanks_01. Important samples should be unique and we have to indicate that in the documentation very clearly.