Updating mmvec installation
I'm not sure if it is exactly correct the way I suggested in the code... but it worked installing mmvec via pip inside the qiime2 conda environment.
When we install mmvec, it automatically will install tensorflow 1.15.4. It will require to have numpy<1.19.0, >=1.16.0. 
Initially, I had 1.19.4 and I downgraded to numpy==1.18.0.
The mmvec workflow ran successfully, and the lodgir folder was generated.